+++ src/game/scripts/vscripts/pregame.lua
@@ -2362,7 +2362,7 @@ function Pregame:initOptionSelector()
 
         -- Other - Extra ability
         lodOptionExtraAbility = function(value)
-            return value == 0 or value == 1 or value == 2 or value == 3 or value == 4  or value == 5 or value == 6 or value == 7 or value == 8
+            return value == 0 or value == 1 or value == 2 or value == 3 or value == 4  or value == 5 or value == 6 or value == 7 or value == 8 or value == 9 or value == 10  or value == 11 or value == 12
         end,
 
         -- Other -- Gotta Go Fast!
@@ -3164,32 +3164,29 @@ function Pregame:processOptions()
         -- All extra ability mutator stuff
         if this.optionStore['lodOptionExtraAbility'] == 1 then
             this:banAbility("gemini_unstable_rift")
-        end
-
-        if this.optionStore['lodOptionExtraAbility'] == 2 then
+        elseif this.optionStore['lodOptionExtraAbility'] == 2 then
             this:banAbility("imba_dazzle_shallow_grave")
-        end
-
-        if this.optionStore['lodOptionExtraAbility'] == 3 then
+        elseif this.optionStore['lodOptionExtraAbility'] == 3 then
             this:banAbility("imba_tower_forest")
-        end
-
-        if this.optionStore['lodOptionExtraAbility'] == 4 then
+        elseif this.optionStore['lodOptionExtraAbility'] == 4 then
             this:banAbility("ebf_rubick_arcane_echo")
             this:banAbility("ebf_rubick_arcane_echo_OP")
-        end
-
-        if this.optionStore['lodOptionExtraAbility'] == 6 then
+        elseif this.optionStore['lodOptionExtraAbility'] == 6 then
             this:banAbility("ursa_fury_swipes")
             this:banAbility("ursa_fury_swipes_lod")
-        end
-
-        if this.optionStore['lodOptionExtraAbility'] == 7 then
+        elseif this.optionStore['lodOptionExtraAbility'] == 7 then
             this:banAbility("spirit_breaker_greater_bash")
-        end
-
-        if this.optionStore['lodOptionExtraAbility'] == 8 then
+        elseif this.optionStore['lodOptionExtraAbility'] == 8 then
             this:banAbility("death_prophet_witchcraft")
+        elseif this.optionStore['lodOptionExtraAbility'] == 9 then
+            this:banAbility("sniper_take_aim")
+        elseif this.optionStore['lodOptionExtraAbility'] == 10 then
+            this:banAbility("aether_range_lod")
+            this:banAbility("aether_range_lod_op")
+        elseif this.optionStore['lodOptionExtraAbility'] == 11 then
+            this:banAbility("alchemist_goblins_greed")
+        elseif this.optionStore['lodOptionExtraAbility'] == 12 then
+            this:banAbility("angel_arena_nether_ritual")
         end
         
         -- Enable Universal Shop
@@ -6119,6 +6116,18 @@ function Pregame:fixSpawningIssues()
                         elseif OptionManager:GetOption('extraAbility') == 8 then
                             local extraAbility = spawnedUnit:AddAbility("death_prophet_witchcraft")
                             extraAbility:SetLevel(4)
+                        elseif OptionManager:GetOption('extraAbility') == 9 then
+                            local extraAbility = spawnedUnit:AddAbility("sniper_take_aim")
+                            extraAbility:SetLevel(4)
+                        elseif OptionManager:GetOption('extraAbility') == 10 then
+                            local extraAbility = spawnedUnit:AddAbility("aether_range_lod")
+                            extraAbility:SetLevel(4)
+                        elseif OptionManager:GetOption('extraAbility') == 11 then
+                            local extraAbility = spawnedUnit:AddAbility("alchemist_goblins_greed")
+                            extraAbility:SetLevel(4)
+                        elseif OptionManager:GetOption('extraAbility') == 12 then
+                            local extraAbility = spawnedUnit:AddAbility("angel_arena_nether_ritual")
+                            extraAbility:SetLevel(3)
                         end
 
                     end, DoUniqueString('addExtra'), RandomInt(1,3) )